Move profile page to the dashboard. Fixes #444
