+++ .rubocop_todo.yml
@@ -40,7 +40,6 @@ Metrics/ModuleLength:
     - 'app/controllers/concerns/hyrax/curation_concern_controller.rb'
     - 'app/controllers/concerns/hyrax/collections_controller_behavior.rb'
     - 'app/controllers/concerns/hyrax/file_sets_controller_behavior.rb'
-    - 'app/controllers/concerns/hyrax/users_controller_behavior.rb'
     - 'app/helpers/hyrax/hyrax_helper_behavior.rb'
     - 'app/models/concerns/hyrax/ability.rb'
     - 'app/models/concerns/hyrax/solr_document_behavior.rb'
@@ -89,7 +88,6 @@ RSpec/ExampleLength:
     - 'spec/controllers/hyrax/generic_works_controller_spec.rb'
     - 'spec/controllers/hyrax/my/highlights_controller_spec.rb'
     - 'spec/controllers/hyrax/transfers_controller_spec.rb'
-    - 'spec/controllers/hyrax/users_controller_spec.rb'
     - 'spec/forms/hyrax/forms/collection_form_spec.rb'
     - 'spec/forms/hyrax/forms/batch_edit_form_spec.rb'
     - 'spec/forms/hyrax/forms/batch_upload_form_spec.rb'
@@ -152,7 +150,6 @@ RSpec/AnyInstance:
     - 'spec/controllers/hyrax/dashboard_controller_spec.rb'
     - 'spec/controllers/hyrax/mailbox_controller_spec.rb'
     - 'spec/controllers/hyrax/stats_controller_spec.rb'
-    - 'spec/controllers/hyrax/users_controller_spec.rb'
     - 'spec/controllers/hyrax/transfers_controller_spec.rb'
     - 'spec/jobs/file_set_attached_event_job_spec.rb'
     - 'spec/jobs/content_update_event_job_spec.rb'
