+++ app/models/proxy_deposit_request.rb
@@ -83,7 +83,7 @@ class ProxyDepositRequest < ActiveRecord::Base
   private
 
     def send_request_transfer_message_as_part_of_create
-      user_link = link_to(sending_user.name, Hyrax::Engine.routes.url_helpers.profile_path(sending_user.user_key))
+      user_link = link_to(sending_user.name, Hyrax::Engine.routes.url_helpers.user_path(sending_user.user_key))
       transfer_link = link_to('transfer requests', Hyrax::Engine.routes.url_helpers.transfers_path)
       message = "#{user_link} wants to transfer a work to you. Review all #{transfer_link}"
       User.batch_user.send_message(receiving_user, message, "Ownership Change Request")
