+++ app/views/hyrax/dashboard/profiles/_trophy_edit.html.erb
