+++ app/views/hyrax/dashboard/profiles/edit.html.erb
@@ -1,0 +1,14 @@
+<% content_for :page_header do %>
+  <h1><span class="fa fa-id-card"></span> Edit Profile</h1>
+<% end %>
+<div class="panel panel-default panel-user">
+  <div class="panel-body">
+    <div class="col-sm-6 profile">
+      <%= render 'edit_primary' %>
+    </div>
+
+    <div class="col-sm-6 profile">
+      <%= render 'edit_secondary' %>
+    </div>
+  </div>
+</div>
