+++ app/views/hyrax/dashboard/profiles/show.html.erb
@@ -1,0 +1,22 @@
+<% content_for :page_header do %>
+  <h1><span class="fa fa-id-card"></span> <%= t("hyrax.admin.sidebar.profile") %></h1>
+  <% if can? :edit, current_user %>
+    <div class="pull-right">
+      <%= link_to hyrax.edit_dashboard_profile_path(@user), class: "btn btn-primary" do %>
+        <%= t("hyrax.edit_profile") %>
+      <% end %>
+    </div>
+  <% end %>
+<% end %>
+
+  <div class="panel panel-default panel-user">
+    <div class="panel-body">
+      <div class="list-group col-md-5 col-sm-8">
+        <%= render "hyrax/users/vitals", user: @user %>
+
+        <div class="list-group-item">
+          <%= render 'hyrax/users/user_info', user: @user %>
+        </div>
+      </div>
+    </div>
+  </div>
