+++ app/views/hyrax/users/_left_sidebar.html.erb
@@ -1,4 +1,4 @@
 <%= render 'hyrax/users/user', user: @user %>
 
 <br />
-<a class="btn btn-primary" href="<%= hyrax.profiles_path %>"><i class="glyphicon glyphicon-globe"></i> View Users</a>
+<a class="btn btn-primary" href="<%= hyrax.users_path %>"><i class="glyphicon glyphicon-globe"></i> View Users</a>
