+++ app/views/hyrax/users/_profile_tabs.html.erb
@@ -2,7 +2,7 @@
         <li class="active"><a href="#contributions"><i class="glyphicon glyphicon-star"></i> <%= I18n.t('hyrax.user_profile.tab_highlighted') %></a></li>
         <li><a href="#activity_log" ><i class="glyphicon glyphicon-calendar"></i> <%= I18n.t('hyrax.user_profile.tab_activity') %></a></li>
       </ul>
-       
+
       <div class="tab-content">
         <%= render 'contributions', presenter: presenter %>
         <%= render 'activity', presenter: presenter %>
