--- app/views/hyrax/users/edit.html.erb
@@ -1,8 +1,0 @@
-<h1 class="col-sm-12">Edit Profile</h1>
-<div class="col-xs-12 col-sm-5 profile">
-  <%= render 'edit_primary' %>
-</div>
-
-<div class="col-xs-12 col-sm-offset-1 col-sm-6 profile">
-  <%= render 'edit_secondary' %>
-</div>
