+++ spec/views/hyrax/dashboard/profiles/edit.html.erb_spec.rb
@@ -1,4 +1,4 @@
-describe 'hyrax/users/edit.html.erb', type: :view do
+describe 'hyrax/dashboard/profiles/edit.html.erb', type: :view do
   let(:user) { stub_model(User, user_key: 'mjg') }
 
   before do
